// src/webparts/expenseForm/components/ExpenseForm.module.scss.d.ts
declare const styles: {
  expenseForm: string;
  container: string;
  header: string;
  formSection: string;
  fileUpload: string;
  fileLabel: string;
  required: string;
  fileInput: string;
  fileName: string;
  buttonSection: string;
  teams: string;
  welcome: string;
  welcomeImage: string;
  links: string;
};

export default styles;